requirements.txt
pydantic
python-jose
fastapi
uvicorn
sqlalchemy
psycopg2
ormar[postgresql]